backend/requirements.txt
Flask==3.1.1
psycopg[binary]==3.2.9
python-dotenv==1.0.1
Flask-Cors==4.0.1
